// app/types/access-management.ts
export type GitAuthType = 'token' | 'ssh' | 'username_password'

export interface GitCredentialData {
  // GitHub/GitLab Token
  token?: string
  
  // SSH密钥
  privateKey?: string
  publicKey?: string
  passphrase?: string
  
  // 用户名密码
  username?: string
  password?: string
  email?: string
}